+++ src/main/resources/script.sql
@@ -1,5 +1,5 @@
 CREATE TABLE `USERS` (
-  `id` bigint PRIMARY KEY AUTO_INCREMENT,
+  `id` integer PRIMARY KEY AUTO_INCREMENT,
   `email` varchar(255) UNIQUE,
   `name` varchar(255),
   `password` varchar(255),
@@ -8,22 +8,22 @@ CREATE TABLE `USERS` (
 );
 
 CREATE TABLE `RENTALS` (
-  `id` bigint PRIMARY KEY AUTO_INCREMENT,
+  `id` integer PRIMARY KEY AUTO_INCREMENT,
   `name` varchar(255),
-  `surface` DECIMAL(10, 2),
-  `price` DECIMAL(10, 2),
+  `surface` numeric,
+  `price` numeric,
   `picture` varchar(255),
   `description` varchar(2000),
-  `owner_id` bigint NOT NULL,
+  `owner_id` integer NOT NULL,
   `created_at` timestamp DEFAULT CURRENT_TIMESTAMP,
   `updated_at` timestamp DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
   FOREIGN KEY (`owner_id`) REFERENCES `USERS` (`id`)
 );
 
 CREATE TABLE `MESSAGES` (
-  `id` bigint PRIMARY KEY AUTO_INCREMENT,
-  `rental_id` bigint,
-  `user_id` bigint,
+  `id` integer PRIMARY KEY AUTO_INCREMENT,
+  `rental_id` integer,
+  `user_id` integer,
   `message` varchar(2000),
   `created_at` timestamp DEFAULT CURRENT_TIMESTAMP,
   `updated_at` timestamp DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
